Order the feed relationship in ManageClass
