+++ FeedStoreChallenge/Helpers/ManagedCache+CoreDataClass.swift
@@ -11,4 +11,6 @@ import Foundation
 import CoreData
 
 @objc(ManagedCache)
-public class ManagedCache: NSManagedObject {}
+public class ManagedCache: NSManagedObject {
+
+}
