+++ FeedStoreChallenge/Helpers/ManagedCache+CoreDataProperties.swift
@@ -1,0 +1,62 @@
+//
+//  ManagedCache+CoreDataProperties.swift
+//  FeedStoreChallenge
+//
+//  Created by Romeo Flauta on 5/15/21.
+//  Copyright © 2021 Essential Developer. All rights reserved.
+//
+//
+
+import Foundation
+import CoreData
+
+
+extension ManagedCache {
+
+    @nonobjc public class func fetchRequest() -> NSFetchRequest<ManagedCache> {
+        return NSFetchRequest<ManagedCache>(entityName: "ManagedCache")
+    }
+
+    @NSManaged public var timestamp: Date?
+    @NSManaged public var feed: NSOrderedSet?
+
+}
+
+// MARK: Generated accessors for feed
+extension ManagedCache {
+
+    @objc(insertObject:inFeedAtIndex:)
+    @NSManaged public func insertIntoFeed(_ value: ManagedFeedImage, at idx: Int)
+
+    @objc(removeObjectFromFeedAtIndex:)
+    @NSManaged public func removeFromFeed(at idx: Int)
+
+    @objc(insertFeed:atIndexes:)
+    @NSManaged public func insertIntoFeed(_ values: [ManagedFeedImage], at indexes: NSIndexSet)
+
+    @objc(removeFeedAtIndexes:)
+    @NSManaged public func removeFromFeed(at indexes: NSIndexSet)
+
+    @objc(replaceObjectInFeedAtIndex:withObject:)
+    @NSManaged public func replaceFeed(at idx: Int, with value: ManagedFeedImage)
+
+    @objc(replaceFeedAtIndexes:withFeed:)
+    @NSManaged public func replaceFeed(at indexes: NSIndexSet, with values: [ManagedFeedImage])
+
+    @objc(addFeedObject:)
+    @NSManaged public func addToFeed(_ value: ManagedFeedImage)
+
+    @objc(removeFeedObject:)
+    @NSManaged public func removeFromFeed(_ value: ManagedFeedImage)
+
+    @objc(addFeed:)
+    @NSManaged public func addToFeed(_ values: NSOrderedSet)
+
+    @objc(removeFeed:)
+    @NSManaged public func removeFromFeed(_ values: NSOrderedSet)
+
+}
+
+extension ManagedCache : Identifiable {
+
+}
